+++ DND/DND/DND/src/DNDGUISwitchSprite3Color2.cpp
@@ -88,7 +88,7 @@ namespace DND
 		return ret;
 	}
 
-	void SwitchSprite3Color2::SetOrder(int order)
+	void SwitchSprite3Color2::SetOrder(float order)
 	{
 		_spr[0]->SetOrder(order);
 		_spr[1]->SetOrder(order);
